renamed USleep files and cleaned up extra working files.
